requirements.txt
Flask
mysql-connector-python
mysql-connector
Flask-Session
flask-talisman
Flask
pandas
mysql-connector-python==8.0.31
python-dotenv>=1.0.0
pyodbc>=4.0.0
pandas>=1.4.0
numpy>=1.22.0
mlxtend>=0.19.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
SQLAlchemy>=1.4.0
pyodbc>=4.0.0

gunicorn==20.1.0
